Contornando problema onde polos poderiam aparecer duplicados

+++ src/test/java/br/jus/trt4/justica_em_numeros_2016/tabelas_cnj/AnalisaAssuntosCNJTest.java
@@ -25,7 +25,7 @@ public class AnalisaAssuntosCNJTest extends AbstractTestCase {
 			assertNotNull(assunto);
 			
 			// Testa os campos do assunto nacional
-			assertEquals(2540, assunto.getCodigoNacional());
+			assertEquals(2540, (int) assunto.getCodigoNacional());
 			assertNull(assunto.getAssuntoLocal());
 		}
 	}
